+++ src/components/Timetracker.js
@@ -1,0 +1,11 @@
+import React, {Component} from 'react'
+
+class Timetracker extends Component {
+  render() {
+    return (
+      <div>Time Left: 157</div>
+    )
+  }
+}
+
+export default Timetracker
