+++ src/components/__test__/App.test.js
@@ -1,0 +1,9 @@
+// import React from 'react'
+// import { mount } from 'enzyme'
+// import App from '../App'
+
+// it('renders without crashing', () => {
+//   const wrapper = mount(<App />)
+//   console.log(wrapper.html())
+//   // expect(wrapper.find(App).to.have.length(1))
+// })
