+++ src/containers/calculator.js
@@ -1,49 +1,7 @@
-import React, {Component} from 'react'
 import {bindActionCreators} from 'redux'
 import {connect} from 'react-redux'
 import {clickOperator, clickOperand} from '../actions/index'
-
-const operands = (propsData) => (
-  propsData.tileNumbers.map(function(number, key){
-    return (
-      <button
-        key={key}
-        onClick={ () => propsData.clickOperand(number, key) }
-        className="button"
-        disabled={ propsData.clickedTiles.indexOf(key)>-1 ? true : false }
-      >
-        {number}
-      </button>
-    )
-  })
-)
-
-class Calculator extends Component {
-  render() {
-    return (
-      <div>
-        <div className="feedback">
-          <div className="hidden" >+10</div>
-        </div>
-        <div className="operands">
-          {
-            operands(this.props)
-          }
-        </div>
-        <div>
-          <div className="operators">
-            <button onClick={ () => this.props.clickOperator("+") } className="button">+</button>
-            <button onClick={ () => this.props.clickOperator("-") } className="button">-</button>
-            <button onClick={ () => this.props.clickOperator("*") } className="button">*</button>
-            <button onClick={ () => this.props.clickOperator("/") } className="button">/</button>
-            <button onClick={ () => this.props.clickOperator("^") } className="button">^</button>
-          </div>
-          <button onClick={ () => this.props.clickOperator("=") } className="button green equals">=</button>
-        </div>
-      </div>
-    )
-  }
-}
+import Calculator from '../components/Calculator'
 
 function mapStateToProps(state) {
   return {
