+++ src/containers/feedback.js
@@ -1,4 +1,3 @@
-import React, {Component} from 'react'
 import {bindActionCreators} from 'redux'
 import {connect} from 'react-redux'
 import {updateFeedback} from '../actions/index'
