+++ src/containers/mainmenu.js
@@ -1,19 +1,8 @@
-import React, {Component} from 'react'
 import {bindActionCreators} from 'redux'
 import {connect} from 'react-redux'
 import {clickPlay} from '../actions/index'
+import MainMenu from '../components/MainMenu'
 
-class MainMenu extends Component {
-	render() {
-		return (
-			<button 
-				onClick={() => this.props.clickPlay()}
-				className="button play green" 
-				disabled={this.props.isPlaying}
-			>Play</button>
-		)
-	}
-}
 
 function mapStateToProps(state) {
 	return {
@@ -24,5 +13,5 @@ function mapStateToProps(state) {
 function matchDispatchToProps(dispatch) {
 	return bindActionCreators({clickPlay: clickPlay}, dispatch)
 }
-   
+
 export default connect(mapStateToProps, matchDispatchToProps)(MainMenu)
